\name{polychor.matrix}
\alias{polychor.matrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Actually, what does this do? Convert a matrix of phi coefficients to a matrix of polycoric correlations }
\description{Given a vector of a vector of frequencies, use John Fox's polycor function to convert these to polychoric correlations.   

Not ready for public consumption.
}
\usage{
polychor.matrix(x, y = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{a matrix of phi coefficients }
  \item{y}{perhaps  }
}
\details{This is a stub of a function to convert a matrix of phi coefficients with an accompanying vector of frequencies into polychoric correlations.  Clearly it is not there yet, but is included in the package as a stopgap.

Please do not use.

}
\value{
}

\author{ William Revelle}

\examples{

}
\keyword{ models }% at least one, from doc/KEYWORDS
\keyword{ models }% __ONLY ONE__ keyword per line
