\name{psycho.demo}
\alias{psycho.demo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Create demo data for psychometrics }
\description{A not very interesting demo of what happens if bivariate continuous data are dichotomized.  Bascially a demo of r, phi, and polychor. }
\usage{
psycho.demo()
}
%- maybe also 'usage' for other objects documented here.
\details{Not one of my more interesting demonstrations.  See 
\url{http://personality-project.org/r/simulating-personality.html} and 
\url{http://personality-project.org/r/r.datageneration.html} for better demonstrations of data generation.
}
\value{a matrix of correlations)
  
}
\references{\url{http://personality-project.org/r/simulating-personality.html} and 
\url{http://personality-project.org/r/r.datageneration.html} for better demonstrations of data generation. }
\author{ William Revelle}

\seealso{\code{\link{VSS.simulate}}}
\examples{

}
\keyword{ multivariate }% at least one, from doc/KEYWORDS
\keyword{ models }% __ONLY ONE__ keyword per line
