\name{small.msq}
\alias{small.msq}
\docType{data}
\title{
A small example data set taken from a larger data set}
\description{
A small subset of 200 cases and 14 variables from the msqR data  set in the psychTools package.  Saved here to allow vignettes to work without using psychTools}
\usage{data("small.msq")}
\format{
  A data frame with 200 observations on the following 14 variables.
  \describe{
    \item{\code{active}}{a numeric vector}
    \item{\code{alert}}{a numeric vector}
    \item{\code{aroused}}{a numeric vector}
    \item{\code{sleepy}}{a numeric vector}
    \item{\code{tired}}{a numeric vector}
    \item{\code{drowsy}}{a numeric vector}
    \item{\code{anxious}}{a numeric vector}
    \item{\code{jittery}}{a numeric vector}
    \item{\code{nervous}}{a numeric vector}
    \item{\code{calm}}{a numeric vector}
    \item{\code{relaxed}}{a numeric vector}
    \item{\code{at.ease}}{a numeric vector}
    \item{\code{gender}}{a numeric vector}
    \item{\code{drug}}{a numeric vector}
  }
}
\details{
See the detailed discussion of the msqR in the psychTools package}

\references{
Revelle, W. and  Anderson, K.J. (1998) Personality, motivation and cognitive performance: Final report to the Army Research Institute on  contract MDA 903-93-K-0008. (\url{https://www.personality-project.org/revelle/publications/ra.ari.98.pdf}).

}
\examples{
data(small.msq)

}
\keyword{datasets}
