\name{multi.hist}
\alias{multi.hist}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{  Multiple histograms on one screen}
\description{Given a matrix or data.frame, produce histograms for each variable in a "matrix" form. 
}
\usage{multi.hist(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ matrix or data.frame}
}

\author{ William Revelle \cr
         Northwestern University \cr
         Evanston, Illinois \cr
 \email{  revelle@northwestern.edu           } \cr
 \url{http://personality-project.org/revelle.html}}

\examples{
multi.hist(attitude) #see graphics window
}
\keyword{ multivariate }% at least one, from doc/KEYWORDS
\keyword{ hplot }% __ONLY ONE__ keyword per line
